requirements.txt
mcp
mysql-connector-python>=9.1.0
sqlalchemy
llama_index
PyYAML
pandas
pymysql